requirements.txt
DeepFace
Flask
Flask-Session
requests
datetime
werkzeug.utils